Move main chat UI to /chat route
